+++ Source/Scene/AttributeType.js
@@ -102,4 +102,34 @@ AttributeType.getMathType = function (attributeType) {
   }
 };
 
+/**
+ * Gets the number of components per attribute.
+ *
+ * @param {AttributeType} attributeType The attribute type.
+ * @returns {Number} The number of components.
+ *
+ * @private
+ */
+AttributeType.getNumberOfComponents = function (attributeType) {
+  switch (attributeType) {
+    case AttributeType.SCALAR:
+      return 1;
+    case AttributeType.VEC2:
+      return 2;
+    case AttributeType.VEC3:
+      return 3;
+    case AttributeType.VEC4:
+    case AttributeType.MAT2:
+      return 4;
+    case AttributeType.MAT3:
+      return 9;
+    case AttributeType.MAT4:
+      return 16;
+    //>>includeStart('debug', pragmas.debug);
+    default:
+      throw new DeveloperError("attributeType is not a valid value.");
+    //>>includeEnd('debug');
+  }
+};
+
 export default Object.freeze(AttributeType);